feat: 사용자 조회, 삭제, 수정 추가

+++ src/main/java/com/jyp/service/my_today_service/controller/UserController.java
@@ -1,7 +1,13 @@
 package com.jyp.service.my_today_service.controller;
 
+import com.jyp.service.my_today_service.dto.ApiResponse;
+import com.jyp.service.my_today_service.dto.UserUpdateDto;
+import com.jyp.service.my_today_service.dto.UserInfoDto;
 import com.jyp.service.my_today_service.model.Users;
 import com.jyp.service.my_today_service.service.UserService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -9,8 +15,8 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
 
+@Tag(name = "유저 API")
 @RestController
-@RequestMapping("/user")
 public class UserController {
     private final UserService userService;
 
@@ -19,30 +25,46 @@ public class UserController {
         this.userService = userService;
     }
 
-    @PostMapping()
-    public ResponseEntity<Users> createUser(@RequestBody Users user) {
-        try {
-            return new ResponseEntity<>(userService.saveUser(user), HttpStatus.OK);
-        } catch (IllegalArgumentException e) {
-            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
-        } catch (Exception e) {
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+    @Operation(summary = "사용자 조회(사용자 ID)")
+    @GetMapping("/v1.0/user/{userId}")
+    public ResponseEntity<ApiResponse<UserInfoDto>> getUserByUserId(@PathVariable String userId) {
+        final Optional<Users> user = userService.findUserByUserId(userId);
+        if (user.isEmpty()) {
+            ApiResponse<UserInfoDto> errorResponse = new ApiResponse<>(false, "USER_NOT_FOUND", "User with ID " + userId + " not found", null);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse); // 404
         }
+        final UserInfoDto userInfoDto = new UserInfoDto(user.get().getUserId(), user.get().getEmail());
+        ApiResponse<UserInfoDto> response = new ApiResponse<>(true, "", "", userInfoDto);
+        return ResponseEntity.ok(response);
     }
 
-    @DeleteMapping
-    public void deleteUser(@PathVariable Long id) {
+    @Operation(summary = "사용자 삭제")
+    @DeleteMapping("/v1.0/user/{id}")
+    public ResponseEntity<ApiResponse<Boolean>> deleteUser(@PathVariable Long id) {
+        final Optional<Users> user = userService.findUserById(id);
+        if (user.isEmpty()) {
+            ApiResponse<Boolean> errorResponse = new ApiResponse<>(false, "USER_NOT_FOUND", "UserId " + id + " not found", false);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
+        }
         userService.deleteUserById(id);
+        ApiResponse<Boolean> response = new ApiResponse<>(true, "", "", true);
+        return ResponseEntity.ok(response);
     }
 
-    @GetMapping("/{userId}")
-    public Optional<Users> getUserByUserId(@PathVariable String userId) {
-        return userService.findUserByUserId(userId);
-    }
 
-    @GetMapping
-    public Optional<Users> getUserById(@PathVariable Long id) {
-        return userService.findUserById(id);
+    @Operation(summary = "사용자 수정")
+    @PutMapping("/v1.0/user/{id}")
+    public ResponseEntity<ApiResponse<UserInfoDto>> updateUser(@PathVariable Long id, @RequestBody @Valid UserUpdateDto userUpdateDto) {
+        final Optional<Users> user = userService.findUserById(id);
+        if (user.isEmpty()) {
+            ApiResponse<UserInfoDto> errorResponse = new ApiResponse<>(false, "USER_NOT_FOUND", "UserId " + id + " not found", null);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
+        }
+        userService.updateUser(id, userUpdateDto.getPassword(), userUpdateDto.getEmail());
+        final UserInfoDto userInfo = new UserInfoDto(user.get().getUserId(), userUpdateDto.getEmail().isEmpty() ? user.get().getEmail() : userUpdateDto.getEmail().get());
+
+        ApiResponse<UserInfoDto> response = new ApiResponse<>(true, "", "", userInfo);
+        return ResponseEntity.ok(response);
     }
 
 }
